+++ client/src/features/admin/adminKeycloak.types.ts
@@ -16,45 +16,19 @@
  * limitations under the License.
  */
 
-export interface ResourcePool {
-  id: number;
-  name: string;
-  classes: ResourceClass[];
-  quota?: Resources;
-  default: boolean;
-  public: boolean;
+export interface KeycloakUser {
+  id: string;
+  firstName: string;
+  lastName: string;
+  email: string;
 }
 
-export interface ResourceClass {
-  id: number;
-  name: string;
-  cpu: number;
-
-  /** Memory (RAM) in Gigabytes */
-  memory: number;
-
-  gpu: number;
-
-  /** Max disk storage in Gigabytes */
-  max_storage: number;
-
-  /** Default disk storage in Gigabytes */
-  default_storage: number;
-
-  default: boolean;
-
-  matching: boolean;
+export interface KeycloakUserQueryParams {
+  realm: string;
+  userId: string;
 }
 
-export interface Resources {
-  cpu: number;
-  memory: number;
-  gpu: number;
-}
-
-export interface ResourcePoolsQueryParams {
-  cpuRequest?: number;
-  gpuRequest?: number;
-  memoryRequest?: number;
-  storageRequest?: number;
+export interface KeycloakUsersQueryParams {
+  realm: string;
+  search?: string;
 }